修改bug、增加安全性

+++ collection-console/src/main/java/com/kochiu/collection/service/SystemService.java
@@ -1,7 +1,10 @@
 package com.kochiu.collection.service;
 
+import cn.hutool.crypto.SecureUtil;
 import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
+import com.kochiu.collection.data.bo.ClearDataBo;
 import com.kochiu.collection.data.bo.PathBo;
+import com.kochiu.collection.data.dto.UserDto;
 import com.kochiu.collection.data.vo.KeyVo;
 import com.kochiu.collection.entity.*;
 import com.kochiu.collection.enums.ErrorCodeEnum;
@@ -11,19 +14,13 @@ import com.kochiu.collection.exception.CollectionException;
 import com.kochiu.collection.properties.SysConfigProperties;
 import com.kochiu.collection.repository.*;
 import com.kochiu.collection.util.DesensitizationUtil;
+import com.kochiu.collection.util.HexUtils;
 import com.kochiu.collection.util.RsaHexUtil;
-import com.kochiu.collection.util.SysUtil;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.RandomStringUtils;
-import org.springframework.beans.factory.support.DefaultListableBeanFactory;
-import org.springframework.boot.web.servlet.MultipartConfigFactory;
-import org.springframework.context.ApplicationContext;
-import org.springframework.context.ConfigurableApplicationContext;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.util.unit.DataSize;
 
-import javax.servlet.MultipartConfigElement;
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.*;
@@ -39,7 +36,6 @@ import static com.kochiu.collection.Constant.RANDOM_CHARS;
 @Service
 public class SystemService {
 
-    private final ApplicationContext applicationContext;
     private final UserCatalogRepository userCatalogRepository;
     private final UserCategoryRepository userCategoryRepository;
     private final UserResourceRepository userResourceRepository;
@@ -69,8 +65,7 @@ public class SystemService {
             "C:\\Recovery", "C:\\Config.Msi", "C:\\MSOCache"
     ));
 
-    public SystemService(ApplicationContext applicationContext,
-                         UserCatalogRepository userCatalogRepository,
+    public SystemService(UserCatalogRepository userCatalogRepository,
                          UserCategoryRepository userCategoryRepository,
                          UserResourceRepository userResourceRepository,
                          SysUserRepository sysUserRepository,
@@ -83,7 +78,6 @@ public class SystemService {
                          SysSecurityRepository securityRepository,
                          SysConfigProperties sysConfigProperties,
                          SysConfigRepository sysConfigRepository) {
-        this.applicationContext = applicationContext;
         this.userCatalogRepository = userCatalogRepository;
         this.userCategoryRepository = userCategoryRepository;
         this.userResourceRepository = userResourceRepository;
@@ -108,6 +102,7 @@ public class SystemService {
     public boolean testServerPath(PathBo pathBo) {
         String path = pathBo.getPath();
         if (path == null || path.trim().isEmpty()) {
+            log.error("导入路径为空");
             return false;
         }
 
@@ -116,11 +111,13 @@ public class SystemService {
         try {
             normalizedPath = Paths.get(path).normalize().toAbsolutePath();
         } catch (Exception e) {
+            log.error("导入路径格式错误");
             return false;
         }
 
         // 检查是否包含敏感目录
         if (isSensitivePath(normalizedPath.toString())) {
+            log.error("导入路径包含敏感目录");
             return false;
         }
 
@@ -134,8 +131,10 @@ public class SystemService {
                     file.delete(); // 删除测试创建的目录
                     return true;
                 }
+                log.error("导入路径不可读写");
                 return false;
             } catch (SecurityException e) {
+                log.error("导入路径不可读写");
                 return false;
             }
         } else {
@@ -151,12 +150,17 @@ public class SystemService {
                         if (created) {
                             testFile.delete();
                         }
+                        else {
+                            log.error("导入路径不可读写");
+                        }
                         return created;
                     }
                     return true;
                 }
+                log.error("导入路径不是目录");
                 return false;
             } catch (Exception e) {
+                log.error("导入路径异常", e);
                 return false;
             }
         }
@@ -200,7 +204,17 @@ public class SystemService {
 
     //  清空数据
     @Transactional(rollbackFor = Exception.class)
-    public void clear() throws CollectionException {
+    public void clearSysData(UserDto userDto, ClearDataBo clearDataBo) throws Exception {
+
+        SysUser user = sysUserRepository.getUser(userDto);
+        //校验密码
+        String password = HexUtils.base64ToHex(clearDataBo.getPassword());
+        password = RsaHexUtil.decrypt(password, securityRepository.getPrivateKey());
+        password = SecureUtil.sha256(password);
+        if(!password.equals(user.getPassword())){
+            throw new CollectionException(ErrorCodeEnum.INVALID_USERNAME_OR_PASSWORD);
+        }
+
         //清空标签
         userResourceTagRepository.remove(null);
         //清空资源
@@ -255,10 +269,6 @@ public class SystemService {
                     public FileVisitResult postVisitDirectory(Path dir, IOException exc) throws IOException {
                         if (exc != null) {
                             throw exc; // 抛出遍历过程中的异常
-                        }
-                        if(!SysUtil.isRunningInDocker()) {
-                            //容器运行不能删映射目录
-                            Files.delete(dir); // 删除目录（此时目录已空）
                         }
                         return FileVisitResult.CONTINUE;
                     }
